修改实体类属性为包装类型

+++ src/dynamic-server/finger-memory-web/src/main/java/work/nobility/fingermemoryweb/mapper/BookCategoryMapper.java
@@ -9,7 +9,7 @@ public interface BookCategoryMapper {
       "value(#{categoryName})")
   @SelectKey(
       keyProperty = "categoryId",
-      resultType = Integer.class,
+      resultType = Long.class,
       before = false,
       statement = "select last_insert_id()"
   )
@@ -21,8 +21,8 @@ public interface BookCategoryMapper {
   Integer updateBookCategory(BookCategory bookCategory);
 
   @Select("select * from fm_book_categorys where category_id = #{id}")
-  BookCategory selectBookCategoryById(Integer id);
+  BookCategory selectBookCategoryById(Long id);
 
   @Delete("delete from fm_book_categorys where category_id = #{id}")
-  Integer deleteBookCategoryById(Integer id);
+  Integer deleteBookCategoryById(Long id);
 }
